Share the Lucene escaping code
